EmailAgent/pyproject.toml
[project]
name = "emailagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "langchain-google-community>=2.0.7",
    "langchain-google-genai>=2.1.2",
    "langgraph>=0.3.30",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
]
